refactor td_encloser
